api: move markdown from shared into api
