+++ packages/api/package.json
@@ -24,6 +24,8 @@
     "./client": "./src/client/index.ts",
     "./client/*": "./src/client/*.ts",
     "./dev/*": "./src/dev/*.ts",
+    "./lib/markdown": "./src/lib/markdown/index.ts",
+    "./lib/markdown/*": "./src/lib/markdown/*.ts",
     "./lib/*": "./src/lib/*.ts",
     "./urbit": "./src/urbit/index.ts",
     "./urbit/*": "./src/urbit/*.ts",
@@ -46,6 +48,7 @@
   "dependencies": {
     "@aws-sdk/client-s3": "^3.190.0",
     "@aws-sdk/s3-request-presigner": "^3.190.0",
+    "@types/mdast": "^4.0.0",
     "@urbit/aura": "^3.0.0",
     "@urbit/nockjs": "^1.6.0",
     "any-ascii": "^0.3.1",
@@ -56,7 +59,13 @@
     "exponential-backoff": "^3.1.1",
     "libphonenumber-js": "^1.11.18",
     "lodash": "^4.17.21",
+    "mdast-util-gfm": "^3.0.0",
+    "mdast-util-to-markdown": "^2.1.2",
+    "remark-gfm": "^4.0.0",
+    "remark-parse": "^11.0.0",
+    "remark-stringify": "^11.0.0",
     "sorted-btree": "^1.8.1",
+    "unified": "^11.0.0",
     "validator": "^13.7.0",
     "zod": "^3.25.76"
   }
