+++ packages/api/src/lib/markdown/index.ts
@@ -20,4 +20,4 @@ export { remarkGroupMentions, parseGroupMentions } from './groupMentionPlugin';
 export type { GroupMention } from './groupMentionPlugin';
 
 // Table extraction (post-processing for message rendering)
-export { convertContent, extractTablesFromContent } from './extractTables';
+export { extractTablesFromContent } from './extractTables';
