+++ packages/api/vitest.config.ts
@@ -3,7 +3,11 @@ import { defineConfig } from 'vitest/config';
 export default defineConfig({
   test: {
     environment: 'node',
-    include: ['src/test/**/*.test.ts', 'src/__tests__/**/*.test.ts'],
+    include: [
+      'src/test/**/*.test.ts',
+      'src/__tests__/**/*.test.ts',
+      'src/lib/**/*.test.ts',
+    ],
     setupFiles: ['./src/test/setup.ts'],
     passWithNoTests: true,
     testTimeout: 30000, // 30s for network operations
